requirements.txt
faiss-cpu
sentence-transformers
PyPDF2
python-docx
pandas
openpyxl
pytesseract
pdf2image
numpy
